# Implement and extend data-specific cartesian index