This PR extends and is a generalization of #1900. Supersedes #1900.

This PR effectively fixes the current issue, where we have uncoalesced reads/writes in our pointwise broadcast expressions.

The fix was to specialize on broadcast expressions where we don't need to use a universal cartesian index for all datalayouts, but instead use a datalayout-specific index (defined as `DataSpecificCartesianIndex`).

Then, we define `getindex`/`setindex` on datalayouts for this index type, and do _not_ swap the indices, as we've launch the kernel with `CartesianIndices` corresponding to `size(parent(data))` (as opposed to `size(data)` for the generalized case).

The result is that all data is accessed in a contiguous / coalesced fashion.