@charleskawczynski 2024-08-15
It turns out that this solution is pretty effective for Float64, but not so much for Float32 ([xref](https://github.com/CliMA/ClimaCore.jl/issues/1910#issuecomment-2291268643))